// src/app/interfaces/productionrecords.ts
export interface IProduct {
    id: number;
    productCode: string;
    productName: string;
    productType: string;
    bgColor: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    inActivated: boolean;
  
}
// ...
export interface IProcess {
    id: number;
    processCode: string;
    processName: string;
    processType: string;
    defaultStandard: string;
    defaultMessage: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    inActivated: boolean;
}



export interface IOperationInstruction {
    id: number;
    process: IProcess;
    rawmaterialName: string;
    product: IProduct;
    usingStandard: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
}

//ShiftSchdule
export interface IShiftSchdule {
    id: number;
    shiftNo: number;
    shiftName: string;
    startShiftDay: number;
    startShiftHour: number;
    startShiftMinute: number;
    endShiftDay: number;
    endShiftHour: number;
    endShiftMinute: number;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
}
// ...
export interface ITeam {
    id: number;
    teamName: string;
    createBy: string;
    createDate: Date;
    updateBy: string;
    updateDate: Date;
    inActivated: boolean;
}
// ...
export interface IProductionOrder {
	productCode: string;
	productName: string;
	orderQty: number;
	fgQty: number;
	deliveryQty: number;
	deliveryMonth: number;
	deliveryYear: number;
	productId: number;
	orderDetail: any[];
}